[test][stdlib] Spin tests for Optional generalizations off into a separate file
